requirements.txt
mlx-whisper>=0.4.0
rumps>=0.4.0
pyaudio>=0.2.14
pyobjc-framework-Quartz>=10.0
pyobjc-framework-Cocoa>=10.0
